requirements.txt
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
Pillow>=9.5.0
opencv-python>=4.8.0
scipy>=1.10.0
scikit-image>=0.21.0
matplotlib>=3.7.0
seaborn>=0.12.0
joblib>=1.3.0
Flask>=2.3.0
tqdm>=4.65.0
requests>=2.31.0